+++ server/src/modules/option_values/option_values.module.ts
@@ -2,9 +2,9 @@ import { Module } from '@nestjs/common';
 import { OptionValuesService } from './option_values.service';
 import { OptionValuesResolver } from './option_values.resolver';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Option } from 'src/options/entities/option.entity';
-import { Product } from 'src/products/entities/product.entity';
-import { SkuValue } from 'src/sku_values/entities/sku_value.entity';
+import { Option } from 'src/modules/options/entities/option.entity';
+import { Product } from 'src/modules/products/entities/product.entity';
+import { SkuValue } from 'src/modules/sku_values/entities/sku_value.entity';
 import { OptionValue } from './entities/option_value.entity';
 
 @Module({
