+++ server/src/modules/options/options.module.ts
@@ -2,12 +2,12 @@ import { Module } from '@nestjs/common';
 import { OptionsService } from './options.service';
 import { OptionsResolver } from './options.resolver';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Product } from 'src/products/entities/product.entity';
-import { OptionValue } from 'src/option_values/entities/option_value.entity';
-import { SkuValue } from 'src/sku_values/entities/sku_value.entity';
+import { Product } from 'src/modules/products/entities/product.entity';
+import { OptionValue } from 'src/modules/option_values/entities/option_value.entity';
+import { SkuValue } from 'src/modules/sku_values/entities/sku_value.entity';
 import { Option } from './entities/option.entity';
 import { ProductImage } from 'src/product_images/entities/product_image.entity';
-import { OptionValuesService } from 'src/option_values/option_values.service';
+import { OptionValuesService } from 'src/modules/option_values/option_values.service';
 
 @Module({
   imports: [
