+++ server/src/modules/product_skus/dto/update-product_skus.input.ts
