+++ server/src/modules/product_skus/entities/product_skus.entity.ts
@@ -1,7 +1,7 @@
 import { ObjectType, Field, ID, Int } from '@nestjs/graphql';
 import { ProductImage } from 'src/product_images/entities/product_image.entity';
-import { Product } from 'src/products/entities/product.entity';
-import { SkuValue } from 'src/sku_values/entities/sku_value.entity';
+import { Product } from 'src/modules/products/entities/product.entity';
+import { SkuValue } from 'src/modules/sku_values/entities/sku_value.entity';
 import {
   Column,
   Entity,
