+++ server/src/modules/products/entities/product.entity.ts
@@ -1,7 +1,52 @@
-import { ObjectType, Field, Int } from '@nestjs/graphql';
+import { ObjectType, Field, ID, Int } from '@nestjs/graphql';
+import { ProductImage } from 'src/product_images/entities/product_image.entity';
+import { Option } from 'src/modules/options/entities/option.entity';
+import { ProductSkus } from 'src/modules/product_skus/entities/product_skus.entity';
+import {
+  Column,
+  CreateDateColumn,
+  Entity,
+  OneToMany,
+  PrimaryGeneratedColumn,
+  UpdateDateColumn,
+} from 'typeorm';
 
 @ObjectType()
+@Entity()
 export class Product {
-  @Field(() => Int, { description: 'Example field (placeholder)' })
-  exampleField: number;
+  @Field(() => ID)
+  @PrimaryGeneratedColumn({ type: 'bigint' })
+  id!: number;
+
+  @Field()
+  @Column()
+  name: string;
+
+  @Field()
+  @Column('longtext')
+  description: string;
+
+  @Field(() => Int)
+  @Column({ type: 'float', default: 0 })
+  price: number;
+
+  @Field(() => [ProductImage])
+  @OneToMany(() => ProductImage, (productImage) => productImage.product)
+  images?: ProductImage[];
+
+  @Field()
+  @CreateDateColumn()
+  createdAt?: Date;
+
+  @Field()
+  @UpdateDateColumn()
+  updatedAt?: Date;
+
+  @Field(() => [Option])
+  @OneToMany(() => Option, (option) => option.product)
+  options?: Option[];
+
+  @Field(() => [ProductSkus])
+  @OneToMany(() => ProductSkus, (productSkus) => productSkus.product)
+  productSkus?: ProductSkus[];
 }
