+++ server/src/product_images/entities/product_image.entity.ts
@@ -1,6 +1,6 @@
-import { ProductSkus } from './../../product_skus/entities/product_skus.entity';
+import { ProductSkus } from './../../modules/product_skus/entities/product_skus.entity';
 import { Field, ID, ObjectType } from '@nestjs/graphql';
-import { Product } from 'src/products/entities/product.entity';
+import { Product } from 'src/modules/products/entities/product.entity';
 import { Column, Entity, ManyToOne, PrimaryGeneratedColumn } from 'typeorm';
 
 @ObjectType()
