--- server/src/products/dto/update-product.input.ts
@@ -1,8 +1,0 @@
-import { CreateProductInput } from './create-product.input';
-import { InputType, Field, Int, PartialType } from '@nestjs/graphql';
-
-@InputType()
-export class UpdateProductInput extends PartialType(CreateProductInput) {
-  @Field(() => Int)
-  id: number;
-}
