--- server/src/products/products.resolver.ts
@@ -1,37 +1,0 @@
-import { Resolver, Query, Mutation, Args, Int } from '@nestjs/graphql';
-import { ProductsService } from './products.service';
-import { Product } from './entities/product.entity';
-import { CreateProductInput } from './dto/create-product.input';
-// import { UpdateProductInput } from './dto/update-product.input';
-
-@Resolver(() => Product)
-export class ProductsResolver {
-  constructor(private readonly productsService: ProductsService) {}
-
-  @Mutation(() => Product)
-  createProduct(
-    @Args('createProductInput') createProductInput: CreateProductInput,
-  ) {
-    return this.productsService.create(createProductInput);
-  }
-
-  @Query(() => [Product], { name: 'products' })
-  findAll() {
-    return this.productsService.findAll();
-  }
-
-  @Query(() => Product, { name: 'product' })
-  findOne(@Args('id', { type: () => Int }) id: number) {
-    return this.productsService.findOne(id);
-  }
-
-  // @Mutation(() => Product)
-  // updateProduct(@Args('updateProductInput') updateProductInput: UpdateProductInput) {
-  //   return this.productsService.update(updateProductInput.id, updateProductInput);
-  // }
-
-  // @Mutation(() => Product)
-  // removeProduct(@Args('id', { type: () => Int }) id: number) {
-  //   return this.productsService.remove(id);
-  // }
-}
